requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
psycopg2-binary
python-dotenv~=1.0.1
pydantic
pydantic[email]
SQLAlchemy~=2.0.31